+++ gradle.properties
@@ -22,4 +22,4 @@ use_access_transformer = false
 
 # Coremod Arguments
 include_mod = true
-coremod_plugin_class_name = srki2k.tweakedpetroleum.core.TweakedPetroleumPlugin
+coremod_plugin_class_name = io.github.srdjanv.tweakedpetroleum.core.tweakedpetroleum.TweakedPetroleumPlugin
