--- src/main/java/srki2k/tweakedpetroleum/api/ihelpers/IPumpjackAddons.java
@@ -1,16 +1,0 @@
-package srki2k.tweakedpetroleum.api.ihelpers;
-
-import srki2k.tweakedpetroleum.api.crafting.TweakedPumpjackHandler;
-
-public interface IPumpjackAddons {
-
-    void initEnergyStorage();
-
-    int[] getReplenishRateAndPumpSpeed();
-
-    TweakedPumpjackHandler.ReservoirContent getChunkContains();
-
-    boolean caseLiquid(int consumed, int pumpSpeed, int oilAmnt);
-
-    boolean caseGas(int consumed, int pumpSpeed, int oilAmnt);
-}
